Completed PaymentManager and added command for swish

+++ src/main/java/io/github/musikhjalpen2022/swishplugin/api/bossan/PaymentData.java
@@ -3,10 +3,10 @@ package io.github.musikhjalpen2022.swishplugin.api.bossan;
 public class PaymentData {
 
     public final String phoneNumber;
-    public final double amount;
+    public final int amount;
     public final String username;
 
-    public PaymentData(String phoneNumber, double amount, String username) {
+    public PaymentData(String phoneNumber, int amount, String username) {
         this.phoneNumber = phoneNumber;
         this.amount = amount;
         this.username = username;
